+++ backend/alpha_project/settings.py
@@ -24,6 +24,7 @@ INSTALLED_APPS = [
     "django.contrib.staticfiles",
     "rest_framework",
     "rest_framework.authtoken",
+    "idps",
     "djoser",
     "django_filters",
     "colorfield",
