+++ backend/employee.csv
@@ -1,0 +1,5 @@
+name,departament,is_director
+Имя1,2,True
+Имя2,2,False
+Имя3,2,False
+Имя4,2,False
