+++ backend/idp.csv
@@ -1,0 +1,4 @@
+title,employee,director,status_idp
+Освоить Java,2,1,в работе
+Освоить C++,3,1,отменен
+Освоить Dart,4,1,не выполнен
