+++ backend/idps/views.py
@@ -1,3 +1,12 @@
-from django.shortcuts import render
+from django.http import HttpResponse
+from rest_framework import permissions, viewsets
 
-# Create your views here.
+from idps.serializers import IdpSerializer, EmployeeSerializer
+from idps.models import Idp
+
+
+class IdpViewSet(viewsets.ModelViewSet):
+    queryset = Idp.objects.all()
+    serializer_class = IdpSerializer
+    permission_classes = [permissions.IsAuthenticatedOrReadOnly]
+    http_method_names = ('get', 'post', 'patch', 'delete')
